# Back port optimization to broadcast_axis to MXNet1.x