@access2rohit 2020-07-24
> Looks good since it's a cherry-pick.
> For functionality check, broadcast_axis specific tests in unix-cpu should verify that on CI.
> For performance check, do you mind pasting comparisons for w/ and w/o this change for 1.x branch just to be sure these perf achievements are valid for 1.x branch as well?

@ChaiBapchya 
the results are identical to the one I posted on my original PR. These results haven't changed over time. Do you still want additional runs ?