@ChaiBapchya 2020-07-24
That PR was against master [2.0]. This PR is for 1.x branch [which forked off from master awhile ago]. Adding the numbers would ensure we do our due diligence before merging this PR.