@access2rohit 2020-07-27
Code | SSD 1 Epoch time (sec) | %age Speedup/Slowdown w.r.t Master (large tensor disabled)
-- | -- | --
Master (large tensor disabled) | 231 | 0% speedup/slowdown
Master (large tensor enabled) | 378 | 63.23% slowdown
Master + CPU Optimized broadcast_axis (large tensor disabled) | 237 | 2.6% slowdown
Master + CPU Optimized broadcast_axis   (large tensor enabled) | 234 | 1.3% slowdown